README: add valid flags to the README
Since the flags that can be part of a metadata.hcl file were only
documented in the integrations repository (private), users couldn't know
which flags were valid, leading to some integrations failing to be
consumed because of that.

To limit this kind of issue, we directly document them in the main
readme, as a sub-section below the paragraph that introduces it.